feat: add custom fields to bypass parameter restrictions of certain distributors, like grok and gemini

+++ migrations/mysql/2025-04-09-084043_model_transform/up.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS model_transform (
-    id INTEGER PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     model_name TEXT NOT NULL,
     map_model_name TEXT NOT NULL,
     is_enabled BOOLEAN NOT NULL DEFAULT true,
     is_deleted BOOLEAN NOT NULL DEFAULT false,
-    created_at INTEGER NOT NULL,
-    updated_at INTEGER NOT NULL
+    created_at BIGINT NOT NULL,
+    updated_at BIGINT NOT NULL
 );
